+++ addon/components/paper-radio-group.js
@@ -11,7 +11,7 @@ export default BaseFocusable.extend({
   // Lifecycle hooks
   didInitAttrs() {
     this._super(...arguments);
-    Ember.assert('{{paper-radio-group}} requires an `onchange` function', this.get('onchange') && typeof this.get('onchange') === 'function');
+    Ember.assert('{{paper-radio-group}} requires an `onChange` function', this.get('onChange') && typeof this.get('onChange') === 'function');
   },
 
   childRadios: computed(function() {
@@ -58,12 +58,12 @@ export default BaseFocusable.extend({
 
     let childRadio = this.get('enabledChildRadios').objectAt(index);
     childRadio.set('focused', true);
-    this.get('onchange')(childRadio.get('value'));
+    this.sendAction('onChange', childRadio.get('value'));
   },
 
   actions: {
-    onchange(value) {
-      this.get('onchange')(value);
+    onChange(value) {
+      this.sendAction('onChange', value);
     }
   }
 });